Moved authenticate method to EndpointClient
