--- e2e/src/main/java/gov/hhs/cdc/trustedintermediary/e2e/AuthClient.java
@@ -1,19 +1,0 @@
-package gov.hhs.cdc.trustedintermediary.e2e;
-
-import java.io.IOException;
-import java.util.Map;
-import org.apache.hc.core5.http.ClassicHttpResponse;
-import org.apache.hc.core5.http.ContentType;
-
-public class AuthClient {
-    private static final String API_ENDPOINT = "/v1/auth";
-
-    public static ClassicHttpResponse authenticate(String clientId, String clientJwt)
-            throws IOException {
-        return HttpClient.post(
-                API_ENDPOINT,
-                "scope=" + clientId + "&client_assertion=" + clientJwt,
-                ContentType.APPLICATION_FORM_URLENCODED,
-                Map.of());
-    }
-}
